+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.2.0 - January 30, 2016
 * Added optional `defaultValue` field to `modify`. Thanks @dashed.
+* Added `isMap` function to check if a value is a map.
 
 ## 2.1.1 - January 27, 2016
 * Getting map size is now constant time.